Waves dependency mostly dropped, lest searchbar
