+++ src/search.js
@@ -1,4 +1,4 @@
-import { mentors, mentorIds } from "./mentors";
+import { mentors } from "./mentors";
 
 const fieldSearch = ({ field, value }, waveIndex) =>
   mentorIds[waveIndex].filter((mentorId) => mentors[mentorId][field] === value);
