--- src/waves.json
@@ -1,13 +1,0 @@
-{
-  "waves": [
-    {
-      "name": "Wave 1"
-    },
-    {
-      "name": "Wave 2"
-    },
-    {
-      "name": "Wave 3"
-    }
-  ]
-}
